gerrit: pass dependencies to runner
By injecting the dependencies down to the runner, it makes it much
easier to test the agent runtime. Currently, the two main dependencies
are the collectors and their associated TTL cache interval value. Both
of which are built from the passed arguments.

The `cache_dir` attribute will be short lived as we move away from
`DataCache` implementation to using `Storage`.

Change-Id: I927cd3d1143798747353e70d27d9ea3295ea8b8a