Remove ancient APIs for registration of custom watolib extensions
Important: This does not concern cmk.gui.plugins.wato, which can not
easily be removed. This is only related to cmk.gui.plugins.watolib.

This was never an public API and there is also no documented and
versioned replacement. Previously, this was used internally for edition
specific features, but we moved away from that approach.

It might be that this was used by some 3rd party extensions of our
web interface. This is now a breaking change for those extensions.
Technically, there is the possibility to move to our new internal API,
but this is not recommended as well, because it is not public and might
change without further notice

Change-Id: I6902c41d29de3db8b483b382d12b2d1cd63020ca