Fix tooltip z index
The radix tooltip was teleported outside of modal, which is why it
didn't inherit the z-index.

We inline the tooltip instead, making sure it doesn't inherit any visual
attributes from the parent.

We move to a layered z-index approach:
  base < title menu < navigation < modal
and within these, you can have
  dropdown < popup < tooltip

This means that when having a rule open with a dropdown and a tooltip:
scrolling them will hide all behind the title menu and show the tooltip
above the dropdown.

CMK-26594

Change-Id: I424bb62d8353697b80354c78ca3db055ec11d110