include libexec in wheel
With this change the libexec files are
included (with the x flag set) in both
the wheels and the pkg_tars.

This is a complated workaround for the
fact that `package_wheel` will not
preserve permissions of the libexec
files as expected.
Once that is the case, this can be
simplified significantly.

Change-Id: I72539c3e5d5a47dae87a3c1eacc8aedb785f996b