agent_based_plugins fixture: prevent precompile side effect
load_and_convert_legacy_checks hard codes the precompile path to the
precompiled_checks_dir.
This causes any tests that use the agent_based_plugins fixture to have a
side effect and create the precompiled_checks_dir path.
With this change we use mocking to point to a temp path instead.

Change-Id: I45ded0bd8338bdd2cae88991134496e78f8d15aa