Fix clustered services for offline node
Hosts tagged with the criticality "Do not monitor" are treated as offline.

If a cluster contained such an offline node, monitoring of that cluster was broken.

Depending on the selected core, different error messages appeared in the web UI:

* CMC: "unpack required a buffer of 4 bytes"
* Nagios: "Failed to lookup IPv4 address of {hostname} via DNS"

These errors occurred because the cluster check attempted to retrieve check results for an offline node.

With this werk, offline cluster nodes are now ignored, just like regular offline hosts.
This prevents the errors and ensures that cluster monitoring functions correctly even when some nodes are offline.
No action is required to benefit from this fix.

SUP-26131

Change-Id: If35f412df2c22a811556b8a8135d4d796d34f016