cisco_meraki: improve ttl cache settings
Use the time span formspec for improved setting of cache ttl value. Fix
ttl values as they should be in seconds not minutes. Added a test that
validates that the output is cached between runs and a help text to
inform the user why they may want to override the TTL value.

Change-Id: Iae05645cd8c38297fe18bbf240527d7902502977