18935 FIX OpenSSL compatibility fix for failure when Baking Agents
An OpenSSL upgrade in supported distributions caused several system commands to fail when executed within the context of an OMD site user.

This cause Baking agents feature to fail, as it relies on these commands.

To resolve this, the following affected commands are now wrapped to ensure they utilize the system-supplied OpenSSL library when invoked by an OMD site user:

* curl
* nslookup
* pdftoppm
* rpmbuild
* scp
* ssh
* zypper

---

This fix applies to the following operating systems:

* SLES 15 SP5, 15 SP6, 15 SP7
* RHEL 9, 10
* Debian 13

Change-Id: I86049e752037a492a2eac7804dfe19a4e0c5e39b