Align metric backend data to RRD data grid (if available)
Our graph rendering currently assumes a fixed grid for all time series in the
same graph. The RRDs are already aligned to the grid of the first RRD series
(which does not necessarily correspond to the grid requested by the frontend,
since the RRDs adjust the grid based on their internal precision, which is
configurable per service). To render data from both backends in the same graph,
we also use the grid of the first RRD series when querying the backend.

Change-Id: I9dd1573a47f63b771ea60b820314d003a2a7b6df