Fix permissions of empty dirs
empty_dirs attribute of pkg_tar sets permissions to 555 instead of 755.
This results in an error, where files can no longer be edited, after
they have been copied.

The default of pkg_mkdirs is 755, which is why we don't need to set the
permissions explicitly. The issue was found on RPMs, because in DEBs the
issue is hidden by altered permissions later in the process. The
alteration of the DEB package should be fixed in a different change.

This regression was introduced in the change with ID:
I89f6b50dc69bb7b3d457c6f56e73547c4d5ccfdf

CMK-27853 CMK-27854

Change-Id: I9499f286b00797d9d1ceef2c89c634878929d33d