Remove daemon_sleep in relay daemon loop
When we push data to the queue it will get processed quickly already
as the loop only redistributes tasks. The queue implementation
in python also allows that currently with the unbounded queue any
`queue.put(...)` does not yield to the event loop. So overall
this is one less config parameter and a snappier system. Win-win.

Change-Id: I3dce6bcdb1700425835860258b4f939d329dccf8