Bugfix in alignment of metrics backend grid to RRD grid
We can only apply a minimum step size if we don't have to align to RRD data
(aka, if we only plot data from the metrics backend). Otherwise, we have to
stick with the RRD data grid. However, note that the RRDs already align to their
internal precision, which is set to 60 seconds by default.

Change-Id: Ic89ddaf6ad7ad31ee60ec8cea26af2deb0dda2c3