Derive python dependencies for relay from package requirements - Approach #3
With the growing number of plugin families which we extract from the
monolithic cmk package, the explicit list of runtime depnendencies in
omd/non-free/relay/BUILD becomes hard to maintain as we have to
replicate each and every dependency from the respective packages another
time.

This change removes the redundancies by deriving the dependencies
from the package requirements.in files for the relay package.

It works like this:

1. Define central data structure PACKAGES, holding references to
   the relevant packaging and requirements.in targets.
2. Compute a requirements.txt from the individual requirements.in files
3. Use requirements.txt to gather dependencies via pip

The requirements.txt computation uses our global requirements.txt as
constraint file to ensure consistent versions across all packages.

Workflow for adding a new plugin family to the relay is:

1. Add the plugin family to PLUGIN_FAMILIES in omd/non-free/relay/BUILD
2. Run "make relock_venv"

Change-Id: I7db3ddc3563c0d1d57e396267873f549a64d23b7