18959 FIX service search: user input deleted
Fix user input deleted in the service search filter name input.

Problem definition:

In the monitor all services filter "Service (regex)" input,
when having services that come alphabetically before the name inputted
by the user but partially matching the input (e.g. "App Memory 1" for a
user input of "Memory"), when the inputted service name has an exact
match (e.g. a service with the exact name "Memory" exists),
the user input name selection will disappear. The first service from the
list is then selected instead.

We don't want to fix this problem by keeping the user input always
visible and selectable, because that results in the service name
appearing twice on the list. Instead, we order the results to show the
exact match first, then choices that start with the user input string,
then the rest.

SUP-26300

Change-Id: Iefea65881072daa756799b4dc820d40855c07286