config_path: pass around the creation context
Returning an object from the contextmanager is
a first step towards ensuring that all actions
that contribute to the creation of a config are
actually performed in the `create` context.

It also removes the need to compute the "previous"
serial from a given one (which requires the serial
to be an integer).
The logic of advancing from one config serial to
the next is now encaplsulated in the `next` method,
and no longer need to be reversible.
In principle, we could use a UUID now.

Change-Id: I452d8a45913f0bd71c7acfade848d6f149b979bb