18879 New 'csm' command-line tool for Checkmk Synthetic Monitoring
The new command-line tool `csm` (Checkmk Synthetic Monitoring) serves as a standalone, lean abstraction layer above the open source tool micromamba.

The background to this is the future discontinuation of the "RCC" functionality that creates environments for executing Robot Framework tests.
In particular, RCC's ability to execute post-installation commands after the environment has been created is not available in "micromamba" (introduced as a replacement for RCC in Checkmk 2.5).

CSM brings back the convenience to create test environments with a single command without having to worry about the details of the micromamba syntax or necessary additional postinstall steps.

The subcommand "csm env" encapsulates all tasks related to environment management (creating, activating, deactivating, and packing/unpacking environments).

The subcommand "csm robot create" is used to quickly initialize new Robot Framework projects. The tool creates a new robot test structure and fills it with a fixed set of skeleton files from a minimal sample project. This provides users with ready-to-use directory and project structures without having to manually copy recurring boilerplate files. The aim is to simplify the start of new projects, ensure consistent structures and significantly reduce the time to the first functional test.

"csm robot" will be given additional subcommands in the future, including the execution of robots with RobotCode profiles (used in VS Code).

The csm tool is available for download on GitHub: https://github.com/Checkmk/csm/releases

CMK-24049

Change-Id: I3445edfe12b772c7bbe9f536252ae0c8c056c9c1