testlib.site: Fix message_broker_port definition
Actually we do need to check the central site version too, since we
can not define the message broker port in 2.3.0, not even for remote
sites. That means if we want to add a remote site connection with an
old release central site (for update tests or mixed release tests),
we need to make sure that the central site is >= 2.4.0 before we set
the message broker port.

Change-Id: Ice7e0fc9b9ae43f1ddfd2328439376a18af48db5