test(plugin_siteless): fix network plugin consistency
We have to mark these two plugins as "offend" until we have all
Manufactorer MIBs we support for network devices. A follow up ticket was
created to add support and unmark this as offending.

Also, needed to fix the entry that was added without the asterisk as
this was accidentally merged in.

CMK-28118

Co-authored-by: Benjamin Knapp <benjamin.knapp@checkmk.com>
Change-Id: I82d87c9985bd18aba83ef913f3fcd9f11dc8d3f2