17759 FIX Fixed removal of non-sticky acknowledgements
Non-sticky acknowledgements of soft non-OK states have been removed even
when the host or service stayed in the same soft non-OK state. This was
inconsistent with the way Nagios works and has been fixed.

Note: If you e.g. add a non-sticky acknowledgement for a soft-WARN state
and the service goes to soft-CRIT, the acknowledgement will get removed.
In other words: You acknowledge only the current state, not a general
"non-OK" state, which is consistent with Nagios.

SUP-26017

Change-Id: I3907730a4b7e779e1278dc087572ed39eecb33cf