search: handle invalid regex for monitoring
When an invalid regex was passed down to the monitoring search engine,
it would raise a user error in the backend which in turn would nullify
the unified results from the other providers.

This was confusing behavior for the user. Now, we are proactively making
sure that the query is a valid regex before passing it down to the
monitoring engine.

We will follow up this change by adding a better hint when searching
monitoring that if a regex pattern is passed, it must be valid.

CMK-27324

Change-Id: I05ee08ac2f3e2859f99df750d7e977fc6eac016f