activate changes: load only the changes that can be activated
Previously we were loading all changes of all the sites passed in
by the caller.  However, these sites could be logged out or disabled.
The response from the rest-api included all the changes of all the
sites. This change ensures we only return the changes that were
activated in the rest-api response.

CMK-27937

Change-Id: Id0711895b2a5d5d1cf2e0928c1d35c5ff5e8a6cd