cisco_meraki: integrate switch port statuses
Main changes to the original plugin:

- heavily relying on pydantic models for validation.
- generated schema based on Meraki API for stronger types.
- converted to new style caching mechanism.
- use a time range in the rulespec to define `timespan` configuration.
- added test coverage for agent and agent based functionality.
- general clean ups to adhere to coding guidelines.

This is currently only fetching switch port statuses from the devices
endpoint. There is a planned follow-up to also enable the ability to
fetch the ports from the organizations resource. This is beneficial
because it uses only one request on the Meraki API instead of N requests
depending on the number of devices.

Note for maintainers: with this plugin it was clear that we can't always
trust the spec coming from Cisco. For instance, the `lldp` and `cdp`
fields were handled as optional in the source MKP. I've continued to
support this by introducing the `PossiblyMissing` annotation for the
pydantic models. This defaults to None when the input value is falsey.

This should be used whenever you see that there are conditional checks
on the data provided by Meraki. I will apply this back to existing
plugins in follow-up changes.

CMK-28336

Co-authored-by: th.l <thl-cmk@outlook.com>
Change-Id: I027a24af8b9560732cd531b6797fbec5fb3e5fd5