test(search): use polyfactory for other tests
It makes it way more clear what is actually important in each of the
test cases. The `build` method on factory generates an object based on
the dataclass spec and then we can simply specify which fields to
override. That way all else stays constant and we are only testing the
differences between object.

Change-Id: I519ba101e428b0e3657d7bfdf30f53f608e97194