vue(search): display context in title
The "context" attribute is essential for differentiating result items
that share the same title. This attribute is provided by the monitoring
engine when there is more than one host present in site.

The display of the context is intentionally unlike the title (dimmed) to
signify users that they cannot search for it.

CMK-28290

Change-Id: I2fe84f11a0f3ab305a45913488e7829b9f592e3b