Clean up cmk base edition specific handling
All available modes were registered globally, depending on the
existence of certain files. This made it hard to understand which
modes were available in which edition.

This commit moves the registration of modes to edition specific app
objects, making it clearer which modes are available in which edition.

The edition specific application object is created during import time
based on the detected edition. The object is passed to the mode, so that
it can also transport information, like the edition to the mode.

In the next step, we can apply the same pattern to the automations.

Change-Id: I409694d15f548d05bf4cdb21c150ddb134ca8fa3