Retrieve downloads through our pull-through cache
The second line per domain acts as a fallback.
In case someone can't use our internal pull-through cache Bazel will
then move on to the next URL and try that.
We point it at the original URL to have it working when not in our
internal network.

Note: On 2.4.0 we use Bazel 7 and have to resort to the CLI parameter
experimental_downloader_config.

References CMK-21869

Change-Id: I9476abab6ef807aff876ccaf0483fd3dc6cb8107
(cherry picked from commit 8c589df7a43954df56565427fd48668fd60445c8)