17669 FIX Deprecated ruleset Microsoft SQL Server (Windows) will be removed in next version
As announced in werk #15844 the rule set is deprecated. With this change it is not possible anymore
to add new rules in this ruleset. Please migrate to the new plugin if not already done. The rule
set in the Agent Bakery is called "Microsoft SQL Server (Linux, Windows)".

The old ruleset will still exist in Checkmk 2.5.0 and all previously configured rules are considered
to allow a smooth transition. The plugin and it's Agent Bakery ruleset will be removed in the next major
release after Checkmk 2.5.0.

CMK-23068

Change-Id: I5b5b7c0b05421db0cad807ffbf7cd415779edb47