18911 FIX activate changes: include only the changes that were activated in the rest-api response
Previously we were loading all changes of all the sites passed in
by the caller.  However, these sites could be logged out or disabled.
The response from the rest-api included all the changes of all the
sites. This change ensures we only return the changes that were
activated in the rest-api response.

CMK-27937

Change-Id: I90fae2f609e7aa4d2823c4f82ed2282de309bba9