cisco_meraki: define section type for all plugins
This brings consistency to the plugin family where we are always
defining the `Section` type at the top of the file. This prevents issues
where the wrong type is passed to a handler. We are also dropping
optional types in checks & agents handlers because there is no need to
handle such cases as they will never be called if `None`.

Change-Id: I03d2d6647f611fd1aa637680346e12c2d0e28d85