test(agent-receiver): remove tests checking mock socket behavior
The removed tests only made sure that our mock socket behaved correctly
under various concurrency scenarios. But the system we want to test is the
agent-receiver. The tests did not verify anything in the agent receiver code.

So let's remove them as those tests would belong to the cmc in the socket
implementation in the end.

Change-Id: I61e7ddc60bbb427b68c7507de23e57b1dea1e69d
JIRA-Ref: SAASDEV-6224