nagios config generation: skip disabled plugins
When determining the needed check plugins for the
nagios host checks, we can omit the disabled ones.

While I can not see a reason why this change should
be a problem, I am slightly worried.
It seems to be an obvious thing to do now, but why
has it not been done like that in the past?

Let's change the flag, but keep it easily revertable
for now.

Change-Id: Iebbb2e8665d34d4e0d0228583729c78d3d6661d6