Updating release of werk 18920 from 2.4.0p18 to 2.4.0p17
The changes for this werk were present in the 2.4.0p17 release, but this
werk file was not included in that release.

This werk will not be visible in the werks browser sites that have upgraded
or installed the 2.4.0p17 release. Though it will appear on the website.
This has been deemed acceptable.

SUP-26442
SUP-26462

Change-Id: I9bfc7fccd0c2c8a6a4ff33cb8fe61616a787879a