Revert "agent deployment status: remove orphaned files after grace period"
This reverts commit 047416aad672d02ce49b177ad17f6a8129e886e1.

Reason: There files in `var/check_mk/agent_deployment/`, not directories. So,
this diskspace will fail with an exception.

This change was introduced to address CMK-15746, the agent updater
leaves files behind, which could be cleaned up. On the other hand, there
is no problem in leaving them be. Moreover, it would be better for the
component that creates the files to also delete them. This way there is
no mismatch between the component logic and the expectations diskspace
has.

CMK-28496

Change-Id: I64a4e08ae167bb13316906ee1da67ee74646741c