Add bazelisk to gh action cache
This should spare us from issues of downloading bazel via bazelisk:
```
2025/12/16 14:02:35 Downloading https://github.com/aspect-build/aspect-cli/releases/download/2025.11.0/bazel-2025.11.0-linux-x86_64...
2025/12/16 14:03:02 could not download Bazel: failed to download bazel: failed to download bazel: HTTP GET https://github.com/aspect-build/aspect-cli/releases/download/2025.11.0/bazel-2025.11.0-linux-x86_64 failed: unable to complete 5 requests to https://github.com/aspect-build/aspect-cli/releases/download/2025.11.0/bazel-2025.11.0-linux-x86_64 within 30s. Most recent failure: HTTP 504
A locking of python requirements is needed, but we're executed in the CI, where this should not be done.
It seems you forgot to commit the new lock file. Regenerate with: make relock_venv
make: *** [Makefile:203: check_python_requirements] Error 1
```
https://github.com/Checkmk/checkmk/actions/runs/20270494926/job/58204715891

Change-Id: Id3f14813719d33bce52e260c52c0456f74280df0