Always render titles from callback suggestions
Query the callback on first render to get title and update the button
label from the info provided when selecting a suggestion.

We lose the possibility to set the model value to anything but null
from outside the component to save on queries to the backend.

Simplify some logic to fix some unselectable handling

CMK-26706

Change-Id: Id590934f755d1d7cfacfccef85ec8936531a3965