cisco_meraki: return raw devices from client
By refactoring the code to have the client return raw values, we can
more easily attach the network name to the `Device` type in a follow up
change.

It was initially baked into the client so that it would be easy to build
the parsed device with the organization attributes. This logic now lives
in the agent.

CMK-28341

Change-Id: Idf16a1c93f303385d4f1408673635e9540a3e89c