OTel: Add cmk-collector version resource attribute
When monitoring the collector itself, no version was provided in the
scopes for the telemetry data, which might introduce
incompatibilities when upgrading the builder-config.yaml.

With this change, we add version for the otelcol by providing dist
version for builder-config.yaml.

if some componets are added or removed from the builder-config.yaml the
version should increase to reflect the change.
CMK-27707

Change-Id: I90d3c1cde889deb5fbf2c88a426e220c2de3f16d