add two fake icons
problem is, that the icon selector list is generated from the
file system, and there the icons are named 2fa, but we can not use that
as a python enum value, as those can not start with two. this is a quick
fix, not sure what the long term solution could look like, especially
because the icons might be part of the configuration.

Change-Id: I7e57d5252983887ca622119c90c697267ab1a2bd