Update rules_lint
Add `--fixes=false` as the default is true, which sets
`@aspect_rules_lint//lint:fix`.  This, in turn, fails in subtle ways.

In particular, `ruff` seems to fix the error _in the sandbox_ and
exits with 0.  This does not fix the error in the repo and does not
fail the CI.

We should see if the situation is improved with aspect cli 2, if not, we
could either fix the bug upstream, or run the linters with `bazel build`
or `bazel test` isntead of `bazel lint`.

CMK-21274

Change-Id: Ib2f7b69cff59edec793d22aa0c71d267e5761db8