activate changes: pass only site configurations that can be activated to start method
After recent changes to return only the changes that have been activated, in the
rest api response, I broke the activate changes manager which was checking the
the changes for all sites. Now I'm only passing the site configurations that
were loaded by the manager to the start method.

CMK-27624

Change-Id: I3e5ea95c3e7e7ca33302e4ff82f02ac72a750d50